Script + first csv
First script that just adds cleaned columns to the original file.

Not all the important columns were treated. The ones that could not be fused together were left like that.

I couldn't treat some columns because of multiple values. I don't know if it was voluntary or not from Nicole, I left a message.

To do :
- catching errors
- treating columns that can't be fused
- create a new data frame with only the clean columns
- maybe create a scripts that create this new df according to the use (only important variable, clean all of them) -> the goal is to stop having hard code in the main function