Output only has selected columns
Added dictionaries for the columns we want to output.

The output now only contains the selected columns and not the original ones.

Added a function to treat boolean features (boxes that would either be ticked or not in the filemakerpro file).

Only needs to select the columns we want in the dictionnary.